requirements.txt
###### Requirements without Version Specifiers ######
requests
pyyaml